User can enter and retrieve student information
User clicks either the enter information button or the retrieve student
information or the close button. The user is then directed to either
another button picking or to a enter student info screen. The info
screen save the information to a txt file and the retrieve button screen
asks for either mass or individual print out. If a user clicks the
individual button they are directed to a button list of students and can
then click on the buttons and then a drawing panel with the student info
pops up thus displaying the information.